dry-programming: shifted helper functions into a separate script
